move functor from segtree.hpp
